+++ src/Equibles.Web/ViewModels/ShortVolume/ShortVolumeListItemViewModel.cs
@@ -4,8 +4,8 @@ public class ShortVolumeListItemViewModel
 {
     public string Ticker { get; set; }
     public string Name { get; set; }
-    public long ShortVolume { get; set; }
-    public long ShortExemptVolume { get; set; }
-    public long TotalVolume { get; set; }
+    public decimal ShortVolume { get; set; }
+    public decimal ShortExemptVolume { get; set; }
+    public decimal TotalVolume { get; set; }
     public double ShortPercent { get; set; }
 }
